Create IntegrationTestCommons CMake target
This splits the `LinuxTracingIntegrationTests` target to allow for a similar
target reusing some components (`ChildProcess`, `IntegrationTestPuppet`) but
testing `OrbitService` as a whole.
It performs some renamings and moves some code around.
The name of the module (`LinuxTracingIntegrationTests`) stays the same for now.

Bug: http://b/206081462

Test: Run `LinuxTracingIntegrationTests`.
