+++ src/LinuxTracingIntegrationTests/IntegrationTestPuppetSharedObject.cpp
