+++ src/LinuxTracingIntegrationTests/LinuxTracingIntegrationTest.cpp
@@ -18,9 +18,11 @@
 #include <thread>
 #include <utility>
 
+#include "IntegrationTestChildProcess.h"
+#include "IntegrationTestPuppet.h"
+#include "IntegrationTestUtils.h"
 #include "LinuxTracing/Tracer.h"
 #include "LinuxTracing/TracerListener.h"
-#include "LinuxTracingIntegrationTestPuppet.h"
 #include "ObjectUtils/Address.h"
 #include "ObjectUtils/ElfFile.h"
 #include "ObjectUtils/LinuxMap.h"
@@ -31,18 +33,8 @@
 #include "capture.pb.h"
 
 namespace orbit_linux_tracing_integration_tests {
+
 namespace {
-
-[[nodiscard]] bool IsRunningAsRoot() { return geteuid() == 0; }
-
-[[nodiscard]] bool CheckIsRunningAsRoot() {
-  if (IsRunningAsRoot()) {
-    return true;
-  }
-
-  ERROR("Root required for this test");
-  return false;
-}
 
 [[nodiscard]] int ReadPerfEventParanoid() {
   auto error_or_content = orbit_base::ReadFileToString("/proc/sys/kernel/perf_event_paranoid");
@@ -85,82 +77,6 @@ namespace {
   ERROR("Stadia instance required for this test (but kernel release is \"%s\")", release);
   return false;
 }
-
-class ChildProcess {
- public:
-  explicit ChildProcess(const std::function<int()>& child_main) {
-    std::array<int, 2> parent_to_child_pipe{};
-    CHECK(pipe(parent_to_child_pipe.data()) == 0);
-
-    std::array<int, 2> child_to_parent_pipe{};
-    CHECK(pipe(child_to_parent_pipe.data()) == 0);
-
-    pid_t child_pid = fork();
-    CHECK(child_pid >= 0);
-    if (child_pid > 0) {
-      // Parent.
-      child_pid_ = child_pid;
-
-      // Close unused ends of the pipes.
-      CHECK(close(parent_to_child_pipe[0]) == 0);
-      CHECK(close(child_to_parent_pipe[1]) == 0);
-
-      reading_fd_ = child_to_parent_pipe[0];
-      writing_fd_ = parent_to_child_pipe[1];
-
-    } else {
-      // Child.
-
-      // Close unused ends of the pipes.
-      CHECK(close(parent_to_child_pipe[1]) == 0);
-      CHECK(close(child_to_parent_pipe[0]) == 0);
-
-      // Redirect reading end of parent_to_child_pipe to stdin and close the pipe's original fd.
-      CHECK(close(STDIN_FILENO) == 0);
-      CHECK(dup2(parent_to_child_pipe[0], STDIN_FILENO) == STDIN_FILENO);
-      CHECK(close(parent_to_child_pipe[0]) == 0);
-
-      // Redirect writing end of child_to_parent_pipe to stdout and close the pipe's original fd.
-      CHECK(close(STDOUT_FILENO) == 0);
-      CHECK(dup2(child_to_parent_pipe[1], STDOUT_FILENO) == STDOUT_FILENO);
-      CHECK(close(child_to_parent_pipe[1]) == 0);
-
-      // Run the child and exit.
-      exit(child_main());
-    }
-  }
-
-  ~ChildProcess() {
-    CHECK(close(reading_fd_) == 0);
-    CHECK(close(writing_fd_) == 0);
-
-    CHECK(waitpid(child_pid_, nullptr, 0) == child_pid_);
-  }
-
-  [[nodiscard]] pid_t GetChildPidNative() const { return child_pid_; }
-
-  void WriteLine(std::string_view str) {
-    std::string string_with_newline = std::string{str}.append("\n");
-    CHECK(write(writing_fd_, string_with_newline.c_str(), string_with_newline.length()) ==
-          static_cast<ssize_t>(string_with_newline.length()));
-  }
-
-  [[nodiscard]] std::string ReadLine() {
-    std::string str;
-    while (true) {
-      char c;
-      CHECK(read(reading_fd_, &c, sizeof(c)) == sizeof(c));
-      if (c == '\n' || c == '\0') break;
-      str.push_back(c);
-    }
-    return str;
-  }
-
- private:
-  pid_t child_pid_ = -1;
-  int reading_fd_ = -1;
-  int writing_fd_ = -1;
-};
 
 class BufferTracerListener : public orbit_linux_tracing::TracerListener {
  public:
@@ -326,7 +242,7 @@ class BufferTracerListener : public orbit_linux_tracing::TracerListener {
 // Let's make the fixture ourselves, which is also advised in http://go/totw/122.
 class LinuxTracingIntegrationTestFixture {
  public:
-  LinuxTracingIntegrationTestFixture() : puppet_{&LinuxTracingIntegrationTestPuppetMain} {}
+  LinuxTracingIntegrationTestFixture() : puppet_{&IntegrationTestPuppetMain} {}
 
   [[nodiscard]] pid_t GetPuppetPidNative() const { return puppet_.GetChildPidNative(); }
   [[nodiscard]] uint32_t GetPuppetPid() const {
@@ -393,7 +309,7 @@ class LinuxTracingIntegrationTestFixture {
   std::optional<BufferTracerListener> listener_ = std::nullopt;
 };
 
-using PuppetConstants = LinuxTracingIntegrationTestPuppetConstants;
+using PuppetConstants = IntegrationTestPuppetConstants;
 
 [[nodiscard]] std::vector<orbit_grpc_protos::ProducerCaptureEvent> TraceAndGetEvents(
     LinuxTracingIntegrationTestFixture* fixture, std::string_view command,
